[BUGFIX] Don't remove mailto: as it is a valid URL
According to the MDN[1], "mailto:" is a valid URL for creating an email
link and can be used like that. That means it is not necessary to unset
the mail-url if its content is just "mailto:".

This also restores default behaviour seen in Version 9.x of rx_shariff,
which means the Integrator does not have to comb through many instances
of this plugin manually to restore the former behaviour.

Links:
[1]:
https://developer.mozilla.org/en-US/docs/Learn/HTML/Introduction_to_HTML/Creating_hyperlinks#E-mail_links